Move Query.Operation into it's own file
